requirements.txt
cvxopt
numpy
osqp
pandas
pykale
pytest
pytorch
scikit-learn
scipy
